"Implemented the function that decides whether to change, or add a new key"
"This is probably the hardest function I've done, because to make this work,
a lot of the other functions had to be Thanos snapped. On the bright side,
this current iteration of the hashMaps is extremely modular and
there is great separation of concerns. Working with it will (hopefully)
be easy"